Fix futures canvas candle loading

+++ rust_sr/server/src/models/candles.rs
@@ -1,15 +1,14 @@
 use chrono::NaiveDateTime;
-use rust_decimal::Decimal;
 use serde::Serialize;
 use sqlx::FromRow;
 #[derive(Debug, Clone, Serialize, FromRow)]
 pub struct Candle {
     pub symbol: String,
     pub date: NaiveDateTime,
-    pub open: Decimal,
-    pub high: Decimal,
-    pub low: Decimal,
-    pub close: Decimal,
+    pub open: f64,
+    pub high: f64,
+    pub low: f64,
+    pub close: f64,
     pub volume: i64,
     pub three_month: Option<bool>,
     pub six_month: Option<bool>,
